+++ .package_dev/compile.R
@@ -22,3 +22,8 @@ devtools::release()
 # CXXFLAGS=-fno-gnu-unique in ~/.R/Makevars
 # or
 # PKG_CXXFLAGS=-fno-gnu-unique in src/Makevars
+
+devtools::load_all()
+devtools::load_all()
+
+devtools::install()